Important milestone achieved: Final core function integration
Completed integration of the last core function; implemented in the class BusyResultsPresent
in Task Manager. This function is now responsible for presenting optimal search results.
Achieved the minimum goal of providing a GUI for the core functionality,
mirroring the capabilities of the original DDG SIMULA'67 program.

For more details, see src\gui\README.md.

Next steps:

- Develop graphical presentations of result graphs.
- Translate all originally Hungarian texts and messages into English.
- Rewrite the SIMULA'67 program's Python mirror code in src\main folder into more elegant Python.
(The code is already in Python, but it lacks elegance.)

Immediate priority:

- Perform a comprehensive revision to ensure the application is environment-independent
and ready for download.